confirm组件

+++ src/views/layout/components/header/header-search/history.vue
@@ -24,6 +24,7 @@ const EMITS_ITEM_CLICK = 'itemClick'
 
 <script setup>
 import { useStore } from 'vuex'
+import { confirm } from '@/libs'
 
 const emits = defineEmits([EMITS_ITEM_CLICK])
 
@@ -33,7 +34,11 @@ const store = useStore()
  * 删除所有记录
  */
 const onDeleteAllClick = () => {
-  store.commit('search/deleteAllHistory')
+  confirm('要删除所有历史记录吗？').then(() => {
+    store.commit('search/deleteAllHistory')
+  }).catch((err) => {
+    console.log(err)
+  })
 }
 
 /**
